+++ src/commands.cpp
@@ -5,6 +5,7 @@
 #include "core/conf.h"
 #include "tasks/tasks.h"
 #include "tools/tools.h"
+#include "utility/threading.h"
 
 namespace mob
 {
