+++ src/main.cpp
@@ -6,6 +6,7 @@
 #include "core/op.h"
 #include "tasks/tasks.h"
 #include "tools/tools.h"
+#include "utility/threading.h"
 
 namespace mob
 {
