+++ src/tasks/translations.cpp
@@ -1,5 +1,6 @@
 #include "pch.h"
 #include "tasks.h"
+#include "../utility/threading.h"
 
 namespace mob
 {
