+++ src/tools/git.cpp
@@ -1,6 +1,7 @@
 #include "pch.h"
 #include "tools.h"
 #include "../core/conf.h"
+#include "../utility/threading.h"
 
 namespace mob
 {
